requirements.txt
discord.py
requests
pytz
python-dotenv